requirements.txt
Flask
requests
google-genai
gunicorn
requests
python-dotenv

